chore: deprecate hasValidPhoneNumberHash
