+++ packages/phone-number-privacy/combiner/package.json
@@ -25,7 +25,7 @@
   },
   "dependencies": {
     "@celo/contractkit": "1.0.0-beta3",
-    "@celo/phone-number-privacy-common": "1.0.23",
+    "@celo/phone-number-privacy-common": "1.0.24-dev",
     "@celo/identity": "1.0.0-beta3",
     "@celo/utils": "1.0.0-beta3",
     "blind-threshold-bls": "https://github.com/celo-org/blind-threshold-bls-wasm#e1e2f8a",
