+++ packages/phone-number-privacy/common/src/interfaces/requests.ts
@@ -13,3 +13,9 @@ export interface GetContactMatchesRequest {
   hashedPhoneNumber: string
   sessionID?: string
 }
+
+export interface GetQuotaRequest {
+  account: string
+  hashedPhoneNumber?: string
+  sessionID?: string
+}
